I feel like I should be waving a dead chicken with this commit.
As was discussed in scala-devel, the pattern matching for Scanners generates very nested if statements. This makes the typechecker work really hard to check these if statements, descending into a deeply nested tree. This can blow the stack.

So, what's the solution? Don't touch Scanners at all. Change the definition format for four character constants in SourceFile.scala. Suddenly the problem goes away and we get a lovely big switch statement. The type checker no longer has to do all that work.

Clearly this is a bug in the pattern matcher. See #1456. But as I don't know when I'll have a fix for that, this works as a temporary workaround. I don't understand why it works yet, but it does. 

Hence the sacrificial chicken. 

All hail Zog, protector of the type checker!



git-svn-id: https://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@16407 5e8d7ff9-d8ef-0310-90f0-a4852d11357a