[This patch submitted by ismael juma - commit message his words, but condensed.]
Fix ticket #1600: Serialization and deserialization of hash-based
collections should not re-use hashCode.  The collection is rebuilt
on deserialization - note that this is not compatible with the previous
serialization format. All @SerialVersionUIDs have been reset to 1.

WeakHashMap is not Serializable and should not be so.
TreeHashMap has not been reintegrated yet.
OpenHashMap has not been updated.  (I think this collection is
flawed and should be removed or reimplemented.)

git-svn-id: https://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@19964 5e8d7ff9-d8ef-0310-90f0-a4852d11357a