`any' operator added for immutable lists
It was added directly in the List class because the class is sealed.
We cannot create a simple trait that expects to be mixed in with a List
because it would create a subclass somewhere else than in the List.scala file.
It also avoids to create a wrapper and delegate _all_ the calls to the
underlying list.